[UI] Wire per-game frame rate limit into the performance dialog
Also change framerate_limit to uint32 rather than having to deal with
supporting uint64 for no good reason

+++ src/xenia/gpu/command_processor.cc
@@ -423,8 +423,7 @@ void CommandProcessor::SetDesiredSwapPostEffect(
 
 void CommandProcessor::ThrottlePresentation() {
   // Host frame rate limiting based on framerate_limit cvar.
-  // This is separate from guest vblank timing (controlled by vsync cvar).
-  const uint64_t framerate_limit = cvars::framerate_limit;
+  const uint32_t framerate_limit = cvars::framerate_limit;
 
   if (framerate_limit == 0) {
     // No host frame limiting
